backend/pyproject.toml
[project]
name = "agentic-react-kit-backend"
version = "0.1.0"
description = "ReAct agent kit API (LangGraph + OpenRouter + pluggable tools)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "httpx>=0.28",
    "python-multipart>=0.0.12",
    "PyJWT[crypto]>=2.9",
    "langgraph>=0.2",
    "google-auth>=2.30",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "mcp>=1.6",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
